Wheelspeed Ver 3.0 (has very good bones needs heavy timing investment)
need to get measurements for kph conversion

need to make changes according to new timing configuration

overall it should work when everything is fixed (hopefully)

Ver 3.1 Should drop Friday February 16, 2024 if timing is figured out

+++ Core/Inc/control.h
@@ -25,7 +25,7 @@
 #define RTD_TRACTIVE_VOLTAGE_ON 4500
 
 typedef struct {
-	uint32_t wheelSpeed[4];		 //each index stores speed of each wheel
+	float wheelSpeed[4];		 //each index stores speed of each wheel
 	int32_t motorControllerTemp; // 10:1 conversion
 	int32_t accumulatorMaxTemp; // 10:1 conversion?
 	int32_t coolantTemp; // 10:1 conversion
